LPS-154909 Move functions to utils dir
